requirements.txt
transformers
torch
scikit-learn
ftfy
regex
numpy
opencv-python
clip @ git+https://github.com/openai/CLIP.git
openai>=1.0.0
tqdm>=4.65.0
pillow>=9.0.0
requests>=2.28.0